Cleaning + renaming
